+++ src/extension.ts
@@ -5,13 +5,13 @@ import { Persist } from './Persist';
 export function activate(context: vscode.ExtensionContext) {
   Helper.init(context);
 
-  let selectMarkFromListDisposable = vscode.commands.registerCommand(
-    'taskmarks.selectMarkFromList',
-    () => {
-      Helper.selectMarkFromList();
-    }
-  );
-  context.subscriptions.push(selectMarkFromListDisposable);
+  // let selectMarkFromListDisposable = vscode.commands.registerCommand(
+  //   'taskmarks.selectMarkFromList',
+  //   () => {
+  //     Helper.selectMarkFromList();
+  //   }
+  // );
+  // context.subscriptions.push(selectMarkFromListDisposable);
 
   let selectTaskDisposable = vscode.commands.registerCommand(
     'taskmarks.selectTask',